requirements.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
Pillow>=10.1.0
trimesh==4.0.5
scikit-image==0.25.2
manifold3d==3.0.1 